requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
deep-translator==1.11.4
Django==5.0.6
django-filter==24.2
django-jazzmin==3.0.0
djangorestframework==3.15.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.15.4
fsspec==2024.6.1
google==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
httplib2==0.22.0
huggingface-hub==0.23.5
idna==3.7
inflection==0.5.1
intel-openmp==2021.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openpyxl==3.1.4
packaging==24.1
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyparsing==3.1.2
pypng==0.20220715.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
regex==2024.5.15
requests==2.32.3
rsa==4.9
safetensors==0.4.3
setuptools==70.0.0
soupsieve==2.5
SpeechRecognition==3.10.4
sqlparse==0.5.0
sympy==1.13.0
tbb==2021.13.0
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
